+++ example/lib/generated_plugin_registrant.dart
@@ -1,0 +1,16 @@
+//
+// Generated file. Do not edit.
+//
+
+// ignore: unused_import
+import 'dart:ui';
+
+import 'package:just_audio/just_audio_web.dart';
+
+import 'package:flutter_web_plugins/flutter_web_plugins.dart';
+
+// ignore: public_member_api_docs
+void registerPlugins(PluginRegistry registry) {
+  JustAudioPlugin.registerWith(registry.registrarFor(JustAudioPlugin));
+  registry.registerMessageHandler();
+}