+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.1.10
+
+* Option to set audio session category on iOS.
+
+## 0.1.9
+
+* Bug fixes.
+
+## 0.1.8
+
+* Reduce distortion at slow speeds on iOS
+
 ## 0.1.7
 
 * Minor bug fixes.